+++ doormax/src/main/java/pf/WallToWestOfTaxi.java
@@ -18,12 +18,12 @@ public class WallToWestOfTaxi extends PropositionalFunction {
 				break;
 			}
 		}
-		int taxiX = (int)taxi.getAttributeVal("xLocation");
-		int taxiY = (int)taxi.getAttributeVal("yLocation");
+		Double taxiX = (Double)taxi.getAttributeVal("xLocation");
+		Double taxiY = (Double)taxi.getAttributeVal("yLocation");
 		for (ObjectInstance o :state.objects.values()) {
 			if (o.objectClass.name.equals("wall")) {
-				if(((int)o.getAttributeVal("yLocation") == taxiY)
-					&& ((int)o.getAttributeVal("xLocation") == taxiX -1))
+				if(((Double)o.getAttributeVal("yLocation") == taxiY)
+					&& ((Double)o.getAttributeVal("xLocation") == taxiX -1))
 				return true;
 			}
 		}
